+++ src/main/java/models/TeacherDebt.java
@@ -1,7 +1,6 @@
 package models;
 
 import javax.persistence.*;
-import java.time.Year;
 
 @Entity
 @Table(name = "teacher_debts")
@@ -9,7 +8,7 @@ public class TeacherDebt {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int id;
+    private Long id;
 
     @Column
     private float amount;
@@ -32,11 +31,11 @@ public class TeacherDebt {
     public TeacherDebt() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
